feat(pagination): further splitting up logic
Pagination -> Pager
new Pagination is similar to Async logic, passes in options and handles rendering items and pager
TODO: update async to use new Pagination
