+++ packages/pagination/Pagination.js
@@ -1,229 +1,177 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import classNames from 'classnames';
+// import classNames from 'classnames';
 
-import {
-  Pagination as RsPagination,
-  PaginationLink,
-  PaginationItem as RsPaginationItem,
-} from 'reactstrap';
-import PaginationItem from './PaginationItem';
+import { warnOnce } from './utils';
 
-const buttonTypeProps = PropTypes.oneOfType([PropTypes.string, PropTypes.bool]);
-const defaultButtonText = {
-  firstBtn: '«« First',
-  prevBtn: '« Prev',
-  nextBtn: 'Next »',
-  lastBtn: 'Last »»',
-};
+import Pager from './Pager';
 
 const propTypes = {
-  pagePadding: PropTypes.number,
   page: PropTypes.number,
-  pageCount: PropTypes.number,
-  totalCount: PropTypes.number,
+  // animate: PropTypes.bool, // maybe do something cool in the future
+  scroll: PropTypes.oneOf(['window', 'list', false]),
+  children: PropTypes.func.isRequired,
   itemsPerPage: PropTypes.number,
-  onPageChange: PropTypes.func.isRequired,
-  prevBtn: buttonTypeProps,
-  nextBtn: buttonTypeProps,
-  firstBtn: buttonTypeProps,
-  lastBtn: buttonTypeProps,
-  size: PropTypes.string,
-  align: PropTypes.oneOf(['start', 'center', 'end', 'between']),
-  simple: PropTypes.bool,
-  unstyled: PropTypes.bool,
-  className: PropTypes.string,
+  onPageChange: PropTypes.func,
+  placement: PropTypes.oneOf(['top', 'bottom', 'both']),
+  hideOnSinglePage: PropTypes.bool,
+  options: PropTypes.array,
+  pageOnlyOptions: PropTypes.bool,
+  pageCount: PropTypes.number,
 };
 
 const defaultProps = {
-  pagePadding: 2,
   itemsPerPage: 10,
-  firstBtn: true,
-  prevBtn: true,
-  nextBtn: true,
-  lastBtn: true,
-  align: 'start',
-  unstyled: true,
-  size: 'sm',
+  placement: 'both',
+  scroll: 'list',
+  // animate: true,
 };
 
 class Pagination extends Component {
-  getPageCount() {
-    let output = this.props.pageCount;
-    if (!output && this.props.totalCount && this.props.itemsPerPage > 0) {
-      output = Math.ceil(this.props.totalCount / this.props.itemsPerPage);
-    }
-    return output;
-  }
-
-  isFirstPage() {
-    return this.props.page <= 1;
-  }
-
-  isLastPage() {
-    const lastPage = this.getPageCount();
-    return lastPage ? this.props.page >= lastPage : false;
-  }
-
-  firstPage = () => {
-    this.props.onPageChange(1);
+  state = {
+    controlled: false,
+    page: 1,
   };
 
-  nextPage = () => {
-    this.props.onPageChange(this.props.page + 1);
-  };
-
-  prevPage = () => {
-    this.props.onPageChange(this.props.page - 1);
-  };
-
-  lastPage = () => {
-    this.props.onPageChange(this.getPageCount());
-  };
-
-  getStartButtons() {
-    const { firstBtn, prevBtn } = this.props;
-    const output = [];
-    const firstBtnText =
-      !firstBtn || typeof firstBtn === 'string'
-        ? firstBtn
-        : defaultButtonText.firstBtn;
-
-    if (firstBtnText) {
-      output.push(
-        <RsPaginationItem disabled={this.isFirstPage()} key="firstBtn">
-          <PaginationLink
-            tag="button"
-            type="button"
-            onClick={this.firstPage}
-            aria-label="First"
-          >
-            {firstBtnText}
-          </PaginationLink>
-        </RsPaginationItem>
-      );
-    }
-
-    const prevBtnText =
-      !prevBtn || typeof prevBtn === 'string'
-        ? prevBtn
-        : defaultButtonText.prevBtn;
-    if (prevBtnText) {
-      output.push(
-        <RsPaginationItem disabled={this.isFirstPage()} key="prevBtn">
-          <PaginationLink
-            previous
-            tag="button"
-            type="button"
-            onClick={this.prevPage}
-          >
-            {prevBtnText}
-          </PaginationLink>
-        </RsPaginationItem>
-      );
-    }
-
-    return output;
-  }
-
-  getPages() {
-    const { page, pagePadding } = this.props;
-    const lastPage = this.getPageCount();
-    // start min/max values are page +- padding
-    let minPage = page - pagePadding;
-    let maxPage = page + pagePadding;
-
-    // if min page is lower than 1, add extra values to maxPage
-    if (minPage < 1) {
-      maxPage += 1 - minPage;
-      minPage = 1;
-    }
-
-    // if maxPage is greater than allowed, add values back to minPage
-    if (maxPage > lastPage) {
-      // don't let min get moved to below 1
-      minPage = Math.max(1, minPage - (maxPage - lastPage));
-      maxPage = lastPage;
-    }
-
-    const output = [];
-    if (maxPage > minPage) {
-      for (let i = minPage; i <= maxPage; i += 1) {
-        output.push(
-          <PaginationItem
-            key={i}
-            onClick={this.props.onPageChange}
-            page={i}
-            active={i === page}
-          />
+  constructor(props) {
+    super(props);
+    if (typeof props.page !== 'undefined') {
+      if (typeof props.onPageChange === 'function') {
+        this.state.controlled = true;
+        this.state.page = props.page;
+      } else {
+        warnOnce(
+          'You provided `page` to Pagination but did not provide a function to handle page changes; `page` is being ignored'
         );
       }
     }
-    return output;
+    this.lastRequest = 0;
+    this.scrollRef = React.createRef();
   }
 
-  getEndButtons() {
-    const { nextBtn, lastBtn } = this.props;
-    const output = [];
-    const nextBtnText =
-      !nextBtn || typeof nextBtn === 'string'
-        ? nextBtn
-        : defaultButtonText.nextBtn;
-    if (nextBtnText) {
-      output.push(
-        <RsPaginationItem disabled={this.isLastPage()} key="nextBtn">
-          <PaginationLink
-            previous
-            tag="button"
-            type="button"
-            onClick={this.nextPage}
-          >
-            {nextBtnText}
-          </PaginationLink>
-        </RsPaginationItem>
-      );
+  componentDidUpdate(prevProps, prevState) {
+    if (this.props.scroll && this.state.page !== prevState.page) {
+      if (this.props.scroll === 'window') {
+        window.document.body.scrollIntoView({
+          behavior: 'smooth',
+          block: 'start',
+          inline: 'nearest',
+        });
+      } else if (this.scrollRef.current) {
+        this.scrollRef.current.scrollIntoView({
+          behavior: 'smooth',
+          block: 'start',
+          inline: 'nearest',
+        });
+      }
     }
-    const lastBtnText =
-      !lastBtn || typeof lastBtn === 'string'
-        ? lastBtn
-        : defaultButtonText.lastBtn;
-    if (lastBtnText) {
-      output.push(
-        <RsPaginationItem disabled={this.isLastPage()} key="lastBtn">
-          <PaginationLink
-            tag="button"
-            type="button"
-            onClick={this.prevPage}
-            aria-label="Last"
-          >
-            {lastBtnText}
-          </PaginationLink>
-        </RsPaginationItem>
-      );
-    }
-
-    return output;
   }
+
+  getPageCount() {
+    if (this.props.pageOnlyOptions) {
+      return this.props.pageCount;
+    }
+    return Math.ceil(this.props.options.length / this.props.itemsPerPage);
+  }
+
+  goToPage = page => {
+    if (this.props.onPageChange) this.props.onPageChange(page);
+    if (!this.state.controlled) {
+      this.setState(prevState => {
+        const lastPage = this.getPageCount(prevState);
+        if (page > 0 && (!lastPage || page <= lastPage)) {
+          return {
+            page,
+          };
+        }
+      });
+    }
+  };
 
   render() {
-    const { size, className, align, unstyled } = this.props;
-    const listClassName = classNames(
-      className,
-      `pagination-${unstyled ? 'un' : ''}styles`,
-      `justify-content-${align}`
-    );
+    const {
+      page,
+      scroll,
+      children,
+      itemsPerPage,
+      onPageChange,
+      placement,
+      hideOnSinglePage,
+      options,
+      pageOnlyOptions,
+      ...pagerProps
+    } = this.props;
+
+    const pageCount = this.getPageCount();
+    const showPagination = !hideOnSinglePage || pageCount > 0;
+
+    const usePage = this.state.page;
+
+    let topPager = false;
+    let bottomPager = false;
+    if (showPagination) {
+      const pager = (
+        <Pager
+          {...pagerProps}
+          page={usePage}
+          pageCount={pageCount}
+          onPageChange={this.goToPage}
+          className="my-3"
+        />
+      );
+      if (placement !== 'bottom') {
+        topPager = <div ref={this.scrollRef}>{pager}</div>;
+      }
+      if (placement !== 'top') {
+        bottomPager = pager;
+      }
+    }
+
+    const items = pageOnlyOptions
+      ? options
+      : options.slice((usePage - 1) * itemsPerPage, usePage * itemsPerPage);
+
     return (
       <React.Fragment>
-        <RsPagination size={size} listClassName={listClassName}>
-          {this.getStartButtons()}
-          {!this.props.simple && this.getPages()}
-          {this.getEndButtons()}
-        </RsPagination>
+        {topPager}
+        {children(items || [], this.state)}
+        {bottomPager}
       </React.Fragment>
     );
   }
 }
+
+Pagination.getDerivedStateFromProps = (nextProps, prevState) => {
+  const state = {};
+  if (!prevState.controlled && typeof nextProps.page !== 'undefined') {
+    if (typeof nextProps.onPageChange === 'function') {
+      state.controlled = true;
+      state.page = nextProps.page;
+      warnOnce(
+        'You have changed Pagination from an uncontrolled component to a controlled component; bad things may happen'
+      );
+      return state;
+    }
+    warnOnce(
+      'You provided `page` to Pagination but did not provide a function to handle page changes; `page` is being ignored'
+    );
+  } else if (
+    prevState.controlled &&
+    (typeof nextProps.page === 'undefined' ||
+      typeof nextProps.onPageChange !== 'function')
+  ) {
+    state.controlled = false;
+    warnOnce(
+      'You have changed Pagination from an controlled component to a uncontrolled component; bad things may happen'
+    );
+    return state;
+  } else if (prevState.controlled && nextProps.page !== prevState.page) {
+    state.page = nextProps.page;
+    return state;
+  }
+  return null;
+};
 
 Pagination.propTypes = propTypes;
 Pagination.defaultProps = defaultProps;
